agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "AI Agent"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "openai>=1.106.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
